Create pushup exercise posture correction tracker
Detect the human body.
Finds the position of the different landmarks on the human body based on the model. (Landmarks are shown in image above).
Finds the angle between joints (depending on which joints you select). For my Push Up program, I chose to find the angle at the elbow, shoulder, and hip as those are vital for proper push-up form.